+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 - Refresh the release process documentation (@karalekas, gh-392).
 - Redesign factories, batch runs in BatchedFactory, fix Qiskit utils tests (@rmlarose, @andreamari, gh-381).
 - Added Tensorflow Quantum executor to docs (@k-m-schultz, gh-348).
+- Fix a collection of small build & docs issues (@karalekas, gh-410).
+
+Thanks to @erkska and @k-m-schultz for their contributions to this release!
 
 ## Version 0.2.0 (October 4th, 2020)