+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix a collection of small build & docs issues (@karalekas, gh-410).
 - Add optimal QPR decomposition for depolarizing noise (@karalekas, gh-371).
 - Add PEC basic implementation assuming a decomposition dictionary is given (@andreamari, gh-373).
+- Make tensorflow requirements optional for docs (@karalekas, gh-417).
 
 Thanks to @erkska and @k-m-schultz for their contributions to this release!